feature: stock get all and create

+++ DotNet8.WebApi.SalesOrder/DotNet8.WebApi.SalesOrder/Models/Stock.cs
@@ -10,7 +10,7 @@ namespace DotNet8.WebApi.SalesOrder.Models
         public Guid stockId { get; set; }
         public string? stockName { get; set; }
         public string? stockDescription { get; set; }
-        public float stockPrice { get; set; }
+        public double stockPrice { get; set; }
         public int stockOnHandQty { get; set; }
     }
 }
